added several components for header, added drop down functionalitty
